Conflicts:
	resources/mediawiki.api/mediawiki.api.js

+++ includes/filerepo/file/LocalFile.php
@@ -633,7 +633,7 @@ class LocalFile extends File {
 
 	/**
 	 * Fix thumbnail files from 1.4 or before, with extreme prejudice
-	 * @TODO: do we still care about this? Perhaps a maintenance script
+	 * @todo : do we still care about this? Perhaps a maintenance script
 	 *        can be made instead. Enabling this code results in a serious
 	 *        RTT regression for wikis without 404 handling.
 	 */
@@ -1006,7 +1006,7 @@ class LocalFile extends File {
 	{
 		$pageText = SpecialUpload::getInitialPageText( $desc, $license, $copyStatus, $source );
 
-		if ( !$this->recordUpload2( $oldver, $desc, $pageText ) ) {
+		if ( !$this->recordUpload2( $oldver, $desc, $pageText, false, $timestamp ) ) {
 			return false;
 		}
 
@@ -1043,7 +1043,7 @@ class LocalFile extends File {
 		if ( !$props ) {
 			wfProfileIn( __METHOD__ . '-getProps' );
 			$props = $this->repo->getFileProps( $this->getVirtualUrl() );
-			wfProfileOut( __METHOD__ . -'getProps' );
+			wfProfileOut( __METHOD__ . '-getProps' );
 		}
 
 		if ( $timestamp === false ) {
@@ -1151,10 +1151,12 @@ class LocalFile extends File {
 		# Add the log entry
 		$log = new LogPage( 'upload' );
 		$action = $reupload ? 'overwrite' : 'upload';
-		$log->addEntry( $action, $descTitle, $comment, array(), $user );
+		$logId = $log->addEntry( $action, $descTitle, $comment, array(), $user );
 
 		wfProfileIn( __METHOD__ . '-edit' );
-		if ( $descTitle->exists() ) {
+		$exists = $descTitle->exists();
+
+		if ( $exists ) {
 			# Create a null revision
 			$latest = $descTitle->getLatestRevID();
 			$nullRevision = Revision::newNullRevision(
@@ -1169,6 +1171,15 @@ class LocalFile extends File {
 				wfRunHooks( 'NewRevisionFromEditComplete', array( $wikiPage, $nullRevision, $latest, $user ) );
 				$wikiPage->updateRevisionOn( $dbw, $nullRevision );
 			}
+		}
+
+		# Commit the transaction now, in case something goes wrong later
+		# The most important thing is that files don't get lost, especially archives
+		# NOTE: once we have support for nested transactions, the commit may be moved
+		#       to after $wikiPage->doEdit has been called.
+		$dbw->commit( __METHOD__ );
+
+		if ( $exists ) {
 			# Invalidate the cache for the description page
 			$descTitle->invalidateCache();
 			$descTitle->purgeSquid();
@@ -1176,13 +1187,19 @@ class LocalFile extends File {
 			# New file; create the description page.
 			# There's already a log entry, so don't make a second RC entry
 			# Squid and file cache for the description page are purged by doEdit.
-			$wikiPage->doEdit( $pageText, $comment, EDIT_NEW | EDIT_SUPPRESS_RC, false, $user );
+			$status = $wikiPage->doEdit( $pageText, $comment, EDIT_NEW | EDIT_SUPPRESS_RC, false, $user );
+
+			if ( isset( $status->value['revision'] ) ) { // XXX; doEdit() uses a transaction
+				$dbw->begin();
+				$dbw->update( 'logging',
+					array( 'log_page' => $status->value['revision']->getPage() ),
+					array( 'log_id' => $logId ),
+					__METHOD__
+				);
+				$dbw->commit(); // commit before anything bad can happen
+			}
 		}
 		wfProfileOut( __METHOD__ . '-edit' );
-
-		# Commit the transaction now, in case something goes wrong later
-		# The most important thing is that files don't get lost, especially archives
-		$dbw->commit( __METHOD__ );
 
 		# Save to cache and purge the squid
 		# We shall not saveToCache before the commit since otherwise
@@ -1456,7 +1473,7 @@ class LocalFile extends File {
 	 */
 	function getDescriptionText() {
 		global $wgParser;
-		$revision = Revision::newFromTitle( $this->title );
+		$revision = Revision::newFromTitle( $this->title, false, Revision::READ_NORMAL );
 		if ( !$revision ) return false;
 		$text = $revision->getText();
 		if ( !$text ) return false;
@@ -1467,9 +1484,17 @@ class LocalFile extends File {
 	/**
 	 * @return string
 	 */
-	function getDescription() {
+	function getDescription( $audience = self::FOR_PUBLIC, User $user = null ) {
 		$this->load();
-		return $this->description;
+		if ( $audience == self::FOR_PUBLIC && $this->isDeleted( self::DELETED_COMMENT ) ) {
+			return '';
+		} elseif ( $audience == self::FOR_THIS_USER
+			&& !$this->userCan( self::DELETED_COMMENT, $user ) )
+		{
+			return '';
+		} else {
+			return $this->description;
+		}
 	}
 
 	/**
@@ -2370,25 +2395,32 @@ class LocalFileMoveBatch {
 		$triplets = $this->getMoveTriplets();
 		$triplets = $this->removeNonexistentFiles( $triplets );
 
-		// Copy the files into their new location
-		$statusMove = $repo->storeBatch( $triplets );
-		wfDebugLog( 'imagemove', "Moved files for {$this->file->getName()}: {$statusMove->successCount} successes, {$statusMove->failCount} failures" );
-		if ( !$statusMove->isGood() ) {
-			wfDebugLog( 'imagemove', "Error in moving files: " . $statusMove->getWikiText() );
-			$this->cleanupTarget( $triplets );
-			$statusMove->ok = false;
-			return $statusMove;
-		}
-
 		$this->file->lock(); // begin
+		// Rename the file versions metadata in the DB.
+		// This implicitly locks the destination file, which avoids race conditions.
+		// If we moved the files from A -> C before DB updates, another process could
+		// move files from B -> C at this point, causing storeBatch() to fail and thus
+		// cleanupTarget() to trigger. It would delete the C files and cause data loss.
 		$statusDb = $this->doDBUpdates();
-		wfDebugLog( 'imagemove', "Renamed {$this->file->getName()} in database: {$statusDb->successCount} successes, {$statusDb->failCount} failures" );
 		if ( !$statusDb->isGood() ) {
 			$this->file->unlockAndRollback();
-			// Something went wrong with the DB updates, so remove the target files
-			$this->cleanupTarget( $triplets );
 			$statusDb->ok = false;
 			return $statusDb;
+		}
+		wfDebugLog( 'imagemove', "Renamed {$this->file->getName()} in database: {$statusDb->successCount} successes, {$statusDb->failCount} failures" );
+
+		// Copy the files into their new location.
+		// If a prior process fataled copying or cleaning up files we tolerate any
+		// of the existing files if they are identical to the ones being stored.
+		$statusMove = $repo->storeBatch( $triplets, FileRepo::OVERWRITE_SAME );
+		wfDebugLog( 'imagemove', "Moved files for {$this->file->getName()}: {$statusMove->successCount} successes, {$statusMove->failCount} failures" );
+		if ( !$statusMove->isGood() ) {
+			// Delete any files copied over (while the destination is still locked)
+			$this->cleanupTarget( $triplets );
+			$this->file->unlockAndRollback(); // unlocks the destination
+			wfDebugLog( 'imagemove', "Error in moving files: " . $statusMove->getWikiText() );
+			$statusMove->ok = false;
+			return $statusMove;
 		}
 		$this->file->unlock(); // done
 
@@ -2506,6 +2538,7 @@ class LocalFileMoveBatch {
 		// Create dest pairs from the triplets
 		$pairs = array();
 		foreach ( $triplets as $triplet ) {
+			// $triplet: (old source virtual URL, dst zone, dest rel)
 			$pairs[] = array( $triplet[1], $triplet[2] );
 		}
